- fixed theme linking in ProgressSliderItem - Started to work on rouring and linking in progressslider
